Refactor PCBPlace.py by splitting it into multiple files
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/cpldcpu/PCBFlow?shareId=XXXX-XXXX-XXXX-XXXX).
